requirements.txt
# UI
streamlit==1.32.2
python-dotenv==1.0.1

# LangChain (stable & compatible)
langchain==0.0.353
langchain-community==0.0.20
langchain-google-genai==0.0.6

# Pydantic (required for older LangChain)
pydantic==1.10.13

# Vector DB
chromadb==0.4.24

# Local embeddings (NO API quota issues)
sentence-transformers==2.2.2
torch>=2.0.0

# Gemini SDK
google-generativeai==0.5.4